Create Space Invader Game v0.10
Found out I placed the alien_spawning function in the wrong location, which is why it was causing me issues of the aliens just clumping up and suddenly Drop-bearing my character every so often.

Finally took it outside of the loop and it fixed some of the issues,

Updated the movement code largely to accommodate the rows and columns it had to go up per level, and each level contained a 10% increase in speed.

There's also a Threeshot/tripleshot powerup icon, which isn't being utilized right now, it's currently just floating up and down, but it was also suppose to flash between two images, which it didn't seem to do. and I need to make the image smaller, cuz that thing is HUGE and an eyesore to behold. 

also updated some things, created a dictionary for the alien's movement, not sure how that works, ChatGPT kind of just sent that through, and I just believe it, because all in all, it is working.

Eventually might add LShift speed boost, update the enemy sprite, add more maybe, or maybe they can shoot back.

Increase the screen size, because 800x600 is teeny.

Oh jeez, I forgot a score board, points, and even a level indicator..... maybe next time.